+++ libs/chain-signatures/contract/src/primitives/key_state.rs
@@ -239,13 +239,13 @@ pub mod tests {
     use crate::primitives::key_state::KeyStateProposal;
     use crate::primitives::participants::tests::assert_candidate_migration;
     use crate::primitives::participants::tests::assert_participant_migration;
+    use crate::primitives::test_utils::gen_account_id;
+    use crate::primitives::test_utils::gen_legacy_initializing_state;
+    use crate::primitives::test_utils::gen_legacy_resharing_state;
+    use crate::primitives::test_utils::gen_legacy_running_state;
+    use crate::primitives::test_utils::min_thrershold;
+    use crate::primitives::test_utils::{gen_key_event_id, gen_pk, gen_threshold_params};
     use crate::primitives::thresholds::DKGThreshold;
-    use crate::state::tests::test_utils::gen_account_id;
-    use crate::state::tests::test_utils::gen_legacy_initializing_state;
-    use crate::state::tests::test_utils::gen_legacy_resharing_state;
-    use crate::state::tests::test_utils::gen_legacy_running_state;
-    use crate::state::tests::test_utils::min_thrershold;
-    use crate::state::tests::test_utils::{gen_key_event_id, gen_pk, gen_threshold_params};
     use near_sdk::test_utils::VMContextBuilder;
     use near_sdk::testing_env;
     use rand::Rng;
