+++ libs/chain-signatures/contract/src/primitives/mod.rs
@@ -2,5 +2,7 @@ pub mod key_state;
 pub mod leader;
 pub mod participants;
 pub mod signature;
+#[cfg(test)]
+pub mod test_utils;
 pub mod thresholds;
 pub mod votes;
