+++ libs/chain-signatures/contract/src/state/mod.rs
@@ -2,8 +2,6 @@ pub mod initializing;
 pub mod key_event;
 pub mod resharing;
 pub mod running;
-#[cfg(test)]
-pub mod tests;
 
 use crate::errors::{Error, InvalidState};
 use crate::primitives::key_state::{KeyEventId, KeyStateProposal};
