--- libs/chain-signatures/contract/src/state/tests/mod.rs
@@ -1,1 +1,0 @@
-pub mod test_utils;
